requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
schedule>=1.2.0
tweepy>=4.14.0
flask>=2.3.0
pillow>=10.4.0
selenium>=4.15.0
webdriver-manager>=4.0.1